Corrections made based on reviewer comments as below.
Reviewer Comments
You need to address the following issues:

The App re-initiates network operation to fetch fresh movie list everytime the screen is rotated. This is not only inefficient but user expects the screen to preserve the data on a screen rotation.
The App crashes when rotated while the AsyncTask is still in progress.
Please refer the code review to resolve the issues. :)